+++ llm_guard/input_scanners/gibberish.py
@@ -15,7 +15,6 @@ DEFAULT_MODEL = Model(
     onnx_path="madhurjindal/autonlp-Gibberish-Detector-492513457",
     onnx_revision="fddf42c3008ad61cc481f90d02dd0712ba1ee2d8",
     onnx_subfolder="onnx",
-    pipeline_kwargs={"truncation": True},
 )