+++ llm_guard/model.py
@@ -27,11 +27,18 @@ class Model:
     onnx_revision: Optional[str] = None
     onnx_subfolder: str = ""
     onnx_filename: str = "model.onnx"
-    onnx_enable_hack: bool = True  # Enable hack to remove token_type_ids from input
     kwargs: Dict = dataclasses.field(default_factory=dict)
-    pipeline_kwargs: Dict = dataclasses.field(
-        default_factory=lambda: {"batch_size": 1, "device": device()}
-    )
+    pipeline_kwargs: Dict = dataclasses.field(default_factory=dict)
+
+    def __post_init__(self):
+        default_pipeline_kwargs = {
+            "max_length": 512,
+            "truncation": True,
+            "batch_size": 1,
+            "device": device(),
+            "return_token_type_ids": False,
+        }
+        self.pipeline_kwargs = {**default_pipeline_kwargs, **self.pipeline_kwargs}
 
     def __str__(self):
         return self.path