+++ llm_guard/transformers_helpers.py
@@ -96,10 +96,6 @@ def get_tokenizer_and_model_for_classification(
 
         return tf_tokenizer, tf_model
 
-    # Hack for some models
-    if model.onnx_enable_hack:
-        tf_tokenizer.model_input_names = ["input_ids", "attention_mask"]
-
     tf_model = _ort_model_for_sequence_classification(model)
 
     return tf_tokenizer, tf_model